A Few Changes
Added .gitignore for a few exceptions
Renamed fileSort.py -> main.py
Changed main.py to now include support for other platforms
(Cross-Platform Support).
The README.md now includes instructions to add a function `FileSort` to
bash/zsh as well as powershell.

The requirements.txt is basically the same, it was changed from LF to
CRLF.
