+++ file_sorter/.gitignore
@@ -1,0 +1,4 @@
+.venv
+__pycache__
+test
+testing
